+++ odoo/local-src/specific_payment_mode/models/depiltech_payment_mode.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+# © 2016 Julien Coux (Camptocamp)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+from openerp import models, fields
+
+
+class DepiltechPaymentMode(models.Model):
+    _name = 'depiltech.payment.mode'
+    _order = 'name'
+
+    name = fields.Char(
+        required=True,
+    )